+++ server/src/main/java/com/linbit/linstor/core/objects/SnapshotDefinition.java
@@ -46,10 +46,12 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.UUID;
@@ -526,6 +528,28 @@ public class SnapshotDefinition extends AbsCoreObj<SnapshotDefinition> implement
                 .collect(Collectors.toList()),
             getSnapshotApis(accCtx)
         );
+    }
+
+    /**
+     * Get the average creation time of all snapshots of this snapDfn
+     *
+     * @return
+     */
+    public @Nullable Date getCreationTime()
+    {
+        checkDeleted();
+        int count = 0;
+        long sum = 0;
+        for (Snapshot snap : snapshotMap.values())
+        {
+            Optional<Date> crtTs = snap.getCreateTimestamp();
+            if (crtTs.isPresent())
+            {
+                sum += crtTs.get().getTime();
+                count++;
+            }
+        }
+        return count == 0 ? null : new Date(sum / count);
     }
 
     @Override